refactor: move filter controls to filter directory
